Extract hash into utils/hash
